Refactor books management part
